% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Check_argument_Group_specific_AUC.R
\name{Check_argument_Group_specific_AUC}
\alias{Check_argument_Group_specific_AUC}
\title{FUNCTION_TITLE}
\usage{
Check_argument_Group_specific_AUC(
  MEM_Pol_group,
  time,
  Groups,
  method,
  Averaged
)
}
\arguments{
\item{MEM_Pol_group}{PARAM_DESCRIPTION}

\item{time}{PARAM_DESCRIPTION}

\item{Groups}{PARAM_DESCRIPTION}

\item{method}{PARAM_DESCRIPTION}

\item{Averaged}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\seealso{
\code{\link[ArgumentCheck]{addError}}
}
\keyword{internal}
