% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeltaAUCpckg-package.R
\docType{package}
\name{DeltaAUCpckg-package}
\alias{DeltaAUCpckg}
\alias{DeltaAUCpckg-package}
\title{DeltaAUCpckg: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Marie Alexandre \email{marie.alexandre@inria.fr}

}
\keyword{internal}
