% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Group_specific_AUC_estimation.R
\name{Group_specific_AUC_estimation}
\alias{Group_specific_AUC_estimation}
\title{FUNCTION_TITLE}
\usage{
Group_specific_AUC_estimation(
  MEM_Pol_group,
  time,
  Groups = NULL,
  method = "trapezoid",
  Averaged = FALSE
)
}
\arguments{
\item{MEM_Pol_group}{PARAM_DESCRIPTION}

\item{time}{PARAM_DESCRIPTION}

\item{Groups}{PARAM_DESCRIPTION, Default: NULL}

\item{method}{PARAM_DESCRIPTION, Default: 'trapezoid'}

\item{Averaged}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\seealso{
\code{\link[splines]{bs}}
}
