% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Group_specific_Var_Delta_AUC_estimation.R
\name{Group_specific_Var_Delta_AUC_estimation}
\alias{Group_specific_Var_Delta_AUC_estimation}
\title{FUNCTION_TITLE}
\usage{
Group_specific_Var_Delta_AUC_estimation(
  MEM_Pol_group,
  Group1,
  Group2,
  time.G1,
  time.G2,
  method = "trapezoid",
  Group.dependence = TRUE,
  Averaged = FALSE
)
}
\arguments{
\item{MEM_Pol_group}{PARAM_DESCRIPTION}

\item{Group1}{PARAM_DESCRIPTION}

\item{Group2}{PARAM_DESCRIPTION}

\item{time.G1}{PARAM_DESCRIPTION}

\item{time.G2}{PARAM_DESCRIPTION}

\item{method}{PARAM_DESCRIPTION, Default: 'trapezoid'}

\item{Group.dependence}{PARAM_DESCRIPTION, Default: TRUE}

\item{Averaged}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[ArgumentCheck]{addError}}
\code{\link[splines]{bs}}
}
