% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HIV_Simu_Dataset_Delta01_cens}
\alias{HIV_Simu_Dataset_Delta01_cens}
\title{HIV RNA Load Simulated Data for Delta AUC = -0.1 With Left-censored data}
\format{
A data frame with 4 columns and 1000 rows:
\tabular{ll}{
\code{id} \tab The subject id. \cr
\code{time} \tab Time of HIV RNA load measurements. \cr
\code{Group} \tab Treatment group. \cr
\code{VL} \tab The HIV RNA load measures. Measures are either equals to their true value (if VL > LOD) or to the LOD value (otherwise). \cr
\code{Cens} \tab The indicator of left-censoring: 1 if VLtrue <= LOD and 0 otherwise. \cr
}
}
\usage{
HIV_Simu_Dataset_Delta01_cens
}
\description{
This dataset gives simulated longitudinal log-transformed HIV RNA load dynamics for two distinct groups of patients who interrupted their antiretroviral treatment at baseline (time=0).
The group 'Group1' was simulated to mimic a control group while the group 'Group2' gathers patients assigned to a treatment group.
Each group accounts for 20 patients and each of them were followed for 24 weeks with one measure every week. HIV RNA load values are left-censored due to a limit of detection (LOD), considering log10(50) as threshold value.
This dataset was simulated to obtain a difference of area under the mean HIV RNA load curves between the two groups (AUC2 - AUC1) equals to -0.1 log10 cp/ml.
}
\details{
A dataset containing simulated data used in the accompanying paper to this package.
}
\examples{
HIV_Simu_Dataset_Delta01_cens
}
\keyword{datasets}
