% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEM_Polynomial_Group_structure.R
\name{MEM_Polynomial_Group_structure}
\alias{MEM_Polynomial_Group_structure}
\title{FUNCTION_TITLE}
\usage{
MEM_Polynomial_Group_structure(
  y,
  x = NULL,
  Group = NULL,
  Id = NULL,
  Cens = NULL,
  marginal_dyn_type = "polynomial",
  ind_dyn_type = "polynomial",
  global_intercept = TRUE,
  group_intercept = FALSE,
  degree_group = 3,
  Adaptive = "none",
  min_knots_group = 2,
  max_knots_group = 2,
  knots_group = NULL,
  df_group = NULL,
  Boundary.knots_group = NULL,
  ind_intercept = FALSE,
  degree_ind = 3,
  min_knots_ind = 2,
  max_knots_ind = 2,
  same_base_group_ind = FALSE,
  knots_ind = NULL,
  df_ind = NULL,
  Boundary.knots_ind = NULL,
  ...
)
}
\arguments{
\item{y}{PARAM_DESCRIPTION}

\item{x}{PARAM_DESCRIPTION, Default: NULL}

\item{Group}{PARAM_DESCRIPTION, Default: NULL}

\item{Id}{PARAM_DESCRIPTION, Default: NULL}

\item{Cens}{PARAM_DESCRIPTION, Default: NULL}

\item{marginal_dyn_type}{PARAM_DESCRIPTION, Default: 'polynomial'}

\item{ind_dyn_type}{PARAM_DESCRIPTION, Default: 'polynomial'}

\item{global_intercept}{PARAM_DESCRIPTION, Default: TRUE}

\item{group_intercept}{PARAM_DESCRIPTION, Default: FALSE}

\item{degree_group}{PARAM_DESCRIPTION, Default: 3}

\item{Adaptive}{PARAM_DESCRIPTION, Default: 'none'}

\item{min_knots_group}{PARAM_DESCRIPTION, Default: 2}

\item{max_knots_group}{PARAM_DESCRIPTION, Default: 2}

\item{knots_group}{PARAM_DESCRIPTION, Default: NULL}

\item{df_group}{PARAM_DESCRIPTION, Default: NULL}

\item{Boundary.knots_group}{PARAM_DESCRIPTION, Default: NULL}

\item{ind_intercept}{PARAM_DESCRIPTION, Default: FALSE}

\item{degree_ind}{PARAM_DESCRIPTION, Default: 3}

\item{min_knots_ind}{PARAM_DESCRIPTION, Default: 2}

\item{max_knots_ind}{PARAM_DESCRIPTION, Default: 2}

\item{same_base_group_ind}{PARAM_DESCRIPTION, Default: FALSE}

\item{knots_ind}{PARAM_DESCRIPTION, Default: NULL}

\item{df_ind}{PARAM_DESCRIPTION, Default: NULL}

\item{Boundary.knots_ind}{PARAM_DESCRIPTION, Default: NULL}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\seealso{
\code{\link[splines]{bs}}
\code{\link[lmec]{lmec}}
}
