% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimal_knot_research.R
\name{Optimal_knot_research}
\alias{Optimal_knot_research}
\title{FUNCTION_TITLE}
\usage{
Optimal_knot_research(data, degree, minknot, maxknot)
}
\arguments{
\item{data}{PARAM_DESCRIPTION}

\item{degree}{PARAM_DESCRIPTION}

\item{minknot}{PARAM_DESCRIPTION}

\item{maxknot}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\seealso{
\code{\link[freeknotsplines]{freeknotfit}}
}
