% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AUC_Lagrange_Cjp_coefficients.R
\name{AUC_Lagrange_Cjp_coefficients}
\alias{AUC_Lagrange_Cjp_coefficients}
\title{FUNCTION_TITLE}
\usage{
AUC_Lagrange_Cjp_coefficients(ind_j, ind_p, t)
}
\arguments{
\item{ind_j}{PARAM_DESCRIPTION}

\item{ind_p}{PARAM_DESCRIPTION}

\item{t}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
