% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AUC_Lagrange_Cjp_coefficients.R
\name{AUC_Lagrange_Cjp_coefficients}
\alias{AUC_Lagrange_Cjp_coefficients}
\title{Time Dependent Coefficients Cjp for AUC Lagrange Interpolation Method}
\usage{
AUC_Lagrange_Cjp_coefficients(ind_j, ind_p, t)
}
\arguments{
\item{ind_j}{a numerical scalar indicating the value of the index j.}

\item{ind_p}{a numerical scalar indicating the value of the index p}

\item{t}{a numerical vector of time points (x-axis coordinates) to consider for the AUC calculation.}
}
\value{
a numerical scalar corresponding to the coefficient Cjp evaluated for j = \code{ind_j} and p = \code{ind_p}.
}
\description{
\loadmathjax This function calculate the time-dependent coefficients Cjp involved in the calculation of the are under the curve when the Lagrange interpolation method is used.
}
\details{
The coefficients \mjseqn{C_{jp}} involved in the calculation of the AUC are defined as
\mjsdeqn{C_{2p} = (t_2-t_1)\prod_{l=0 ;\ l\neq p}^{P=2} t_{1+l} - \frac{(t_2^2-t_1^2)}{2}\sum_{l=0 ;\ l\neq p}^{P=2} t_{1+l} + \frac{(t_2^3-t_1^3)}{3} }
\mjsdeqn{C_{mp} = (t_m-t_{m-1})\prod_{l=0 ;\ l\neq p}^{P=2} t_{m-2+l} - \frac{(t_m^2-t_{m-1}^2)}{2}\sum_{l=0 ;\ l\neq p}^{P=2} t_{m-2+l} + \frac{(t_m^3-t_{m-1}^3)}{3} }
\mjsdeqn{C_{jp} = -(t_j-t_{j-1})\prod_{l=0 ;\ l\neq p}^{P=3} t_{j-2+l} + \frac{(t_j^2-t_{j-1}^2)}{2}\sum_{l_1=0 ;\ l_1\neq p}^{P-1=2}\sum_{l_2=l_1+1 ;\ l_2\neq p}^{P=3}t_{j-2+l_1}\cdot t_{j-2+l_2} - \frac{(t_j^3-t_{j-1}^3)}{3} \sum_{l=0 ;\ l\neq p}^{P=3} t_{j-2+l} + \frac{(t_j^4-t_{j-1}^4)}{4}}
where \mjseqn{m} is the number of time points in the vector \code{t}.
}
