% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AUC_Spline_matrix_A.R
\name{AUC_Spline_matrix_A}
\alias{AUC_Spline_matrix_A}
\title{FUNCTION_TITLE}
\usage{
AUC_Spline_matrix_A(time)
}
\arguments{
\item{time}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
