% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AUC_time_weights_estimation.R
\name{AUC_time_weights_estimation}
\alias{AUC_time_weights_estimation}
\title{FUNCTION_TITLE}
\usage{
AUC_time_weights_estimation(time, method)
}
\arguments{
\item{time}{PARAM_DESCRIPTION}

\item{method}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
