% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeltaAUCpckg-package.R
\docType{package}
\name{DeltaAUCpckg-package}
\alias{DeltaAUCpckg}
\alias{DeltaAUCpckg-package}
\title{DeltaAUCpckg: Statistical test - Between group comparison of AUC}
\description{
The goal of DeltaAUCpckg is to propose a statistical test evaluating the difference of Area under the curve (AUC) of a given outcome between two distinct groups of individuals. To this end, longitudinal data obtained for subjects splitted into G distincts groups are fitted with a Mixed-Effects model whose fixed-effects (marginal dynamics) and random-effects (individual dynamics) are respectively modeled by group-structured polynomial or B-splines curves and individual polynomial or B-spline curves.
}
\author{
\strong{Maintainer}: Marie Alexandre \email{marie.alexandre@inria.fr}

}
\keyword{internal}
