% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimal_knot_research.R
\name{Optimal_knot_research}
\alias{Optimal_knot_research}
\title{Optimal Research B-Spline Internal Knots}
\usage{
Optimal_knot_research(data, degree = 3, minknot = 2, maxknot = 2, ...)
}
\arguments{
\item{data}{a data frame containing the independent variable (\emph{y}) and the dependent one (\emph{x}).}

\item{degree}{an integer scalar indicating the degree of the spline it. By default, this value is fixed at 3}

\item{minknot}{an integer scalar indicating the minimum number of knots to consider. By default, this variable is fixed at 2}

\item{maxknot}{an integer scalar indicating the maximum number of knots to consider. By default, this variable is fixed at 2.}

\item{...}{Further arguments to be passed (see \link[freeknotsplines]{freeknotfit} for more details).}
}
\value{
A numerical vector of optimal knots whose number can varied from \code{minknot} to \code{maxknot}
}
\description{
This function uses AIC criterion to estimate optimal position and number of knots by fitting free-knot splines to data with one independent and one dependent variable. It is assumed that knots are estimated for least-squares splines with no penalty using genetic algorithm.
}
\seealso{
\code{\link[freeknotsplines]{freeknotfit}} for more details about the method.
}
